Rework use of yaml dist when testing
Because we don't instrument module-scoped functions, it's hard to
understand how the `yaml` dist is being used when testing shallow
tracking. The call to `yaml.dump` doesn't show up in the expected
appmap. Instead we only see the first class-scoped function it calls,
`Serializer.open`.

Add a new function to `ExampleClass` that uses `yaml.Dumper` directly.