Vendor wrapt, to make sure we use the correct version.
Other projects (e.g. pylint) also use wrapt, and may depend on an older
version. By adding it as a submodule, and rearranging the package
structure slightly, we can ensure that we'll always use the correct
version.

This approach to vendoring was inspired by the way pip does it:
https://pip.pypa.io/en/stable/development/vendoring-policy/, though we
don't need to be as strict. We may want to consider vendoring other
dependencies at some point, too.