Module-scoped functions are now recorded
With these changes, functions defined outside of a class, i.e. with
module scope, are now recorded. In the AppMap, they'll appear as static
functions on a "class".

These changes use wrapt instead of functools.wraps when instrumenting
functions. wrapt does a better job of presenting a wrapped function as
if it was the original function.