fix: Loosen version for some dependencies
Loosen the version specification for a couple of dependencies. Caret
dependencies in pyproject.toml that start with a zero behave differently
than those that start with a non-zero number (see the poetry doc for
details).

In particular, the specification for inflection is too restrictive:
"^0.3.0" will only update to "<0.4.0", and "0.5.1" has been released.