fix: 3.10 support, wrap exec_module properly
wrapt has a problem with methods decorated with @staticmethod, which
results in the arguments to such a method being munged.

In 3.9, importlib.machinery.BuiltinImporter.exec_module was a class
method. It was changed to a static method in 3.10. As a result, trying
to wrap it ran afoul of the shortcoming in wrapt. This caused failures
in various circumstances when mapping: the tests in
test_test_frameworks.py failed, tests in
https://github.com/saleor/saleor failed, etc, etc.

These changes also update the tox config to test 3.10.