feat: Don't show noise in the console.
Show a warning when the agent does something the user is not
necessarily expecting, or when the user made a mistake. So, if
APPMAP=true or APPMAP=false or one of the APPMAP_RECORD_ variables is
set to true or false, log at "info".

If any of those variables is set to a value other than true or false,
log at "warning". This is new behavior.

Make the comparison with true or false be case-insensitive. Before it
was done with '== "true"' and '== "false"'.

If there's an environment variable that starts with the prefix
APPMAP_RECORD_ but doesn't match one of the supported recording types,
log a message at "warning". This is also new behavior.

When recording by default log at "info": it's presumably exactly the
behavior the user was hoping for when they installed the agent, and
they'll see all the AppMaps that got generated.