feat: record unittest testcases by default
No additional configuration is required to record the tests cases in a
subclass of unittest.TestCase. When the tests are run, AppMaps will be
generated.

These changes also improve the way we detect the environment we're
running in. In particular, we ensure that remote recording is not
enabled by default when a Django or Flask app is deployed in production.