fix: improve handling of unset APPMAP
Prior to these changes, the agent still wasn't doing a very good job of
letting the user what was going to happen when APPMAP was unset. Now it
does.

These changes also clean up the tests that check how APPMAP is handled
(set to "true", set to "false", and unset).