fix: ask pytest not to rewrite our modules
Make sure that when we get loaded as a pytest plugin, pytest doesn't try
to rewrite our modules. This avoids a warning about being unable to do
the rewrite, which will cause test failures if warnings have been
promoted to errors (e.g. with warnings.simplefilter("error")).

As far as I can tell, this change doesn't affect our internal tests, and
we still get proper error messages when one of our assertions fails.
Even if this turns out not be true, though, pytest rewriting is cosmetic
only: an assertion failure still raises an exception and the test will
fail. It just won't have a pretty message.