fix: use an RLock in SharedRecorder._add_event
It's possible that when adding an event to the global recorder, the
current thread will try to add another event. This happens in Django's
tests, because it uses weak references with instrumented finalizers.
Calling _add_event can cause the weak reference to get garbage
collected, resulting in a call to the finalizer, which will also call
_add_event.