test: make sure JSON serialization works
With the previous function in simple.py, the return event generated to
check that numpy.int64 serialized correctly didn't test AppMapEncoder.
Keys in a dict that are in a return event are rendered with "repr"
before the event gets serialized. "repr(int64(0))" returned "0" in 1.x,
but returns "numpy.int64(0)" in 2.x, which caused the test to fail.